requirements.txt
langchain-core
langchain-openai
langchain-community
openai
faiss-cpu
python-dotenv
fastapi
uvicorn
edge-tts
